refactor ( #23 ) : UseCase custom annotation infrastructure 모듈로 이동
