+++ casper-application-infrastructure/src/main/kotlin/hs/kr/entrydsm/application/global/aop/GlobalComponentScan.kt
@@ -1,6 +1,6 @@
 package hs.kr.entrydsm.application.global.aop
 
-import hs.kr.entrydsm.global.annotation.usecase.UseCase
+import hs.kr.entrydsm.application.global.annotation.usecase.UseCase
 import org.springframework.context.annotation.ComponentScan
 import org.springframework.context.annotation.ComponentScan.Filter
 import org.springframework.context.annotation.Configuration
